Move C++ classes that implement label matrices to input_data.h/cpp and the corresponding Cython wrappers to input_data.pxd/pyx.
