+++ python/boomer/boosting/cpp/label_wise_losses.cpp
@@ -9,7 +9,7 @@ AbstractLabelWiseLoss::~AbstractLabelWiseLoss() {
 }
 
 std::pair<float64, float64> AbstractLabelWiseLoss::calculateGradientAndHessian(
-        statistics::AbstractLabelMatrix* labelMatrix,  intp exampleIndex, intp labelIndex, float64 predictedScore) {
+        input::AbstractLabelMatrix* labelMatrix,  intp exampleIndex, intp labelIndex, float64 predictedScore) {
     return std::make_pair(0, 0);
 }
 
@@ -18,7 +18,7 @@ LabelWiseLogisticLossImpl::~LabelWiseLogisticLossImpl() {
 }
 
 std::pair<float64, float64> LabelWiseLogisticLossImpl::calculateGradientAndHessian(
-        statistics::AbstractLabelMatrix* labelMatrix, intp exampleIndex, intp labelIndex, float64 predictedScore) {
+        input::AbstractLabelMatrix* labelMatrix, intp exampleIndex, intp labelIndex, float64 predictedScore) {
     uint8 trueLabel = labelMatrix->getLabel(exampleIndex, labelIndex);
     float64 expectedScore = trueLabel ? 1 : -1;
     float64 exponential = exp(expectedScore * predictedScore);
@@ -32,7 +32,7 @@ LabelWiseSquaredErrorLossImpl::~LabelWiseSquaredErrorLossImpl() {
 }
 
 std::pair<float64, float64> LabelWiseSquaredErrorLossImpl::calculateGradientAndHessian(
-        statistics::AbstractLabelMatrix* labelMatrix, intp exampleIndex, intp labelIndex, float64 predictedScore) {
+        input::AbstractLabelMatrix* labelMatrix, intp exampleIndex, intp labelIndex, float64 predictedScore) {
     uint8 trueLabel = labelMatrix->getLabel(exampleIndex, labelIndex);
     float64 expectedScore = trueLabel ? 1 : -1;
     float64 gradient = (2 * predictedScore) - (2 * expectedScore);
