+++ python/boomer/boosting/cpp/label_wise_rule_evaluation.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "../../common/cpp/arrays.h"
-#include "label_wise_losses.h"
+#include "../../common/cpp/rule_evaluation.h"
 
 
 namespace rule_evaluation {
