+++ python/boomer/common/sequential_rule_induction.pxd
@@ -2,7 +2,7 @@ from boomer.common._arrays cimport uint32, intp
 from boomer.common.rules cimport RuleModel, ModelBuilder
 from boomer.common.rule_induction cimport FeatureMatrix, RuleInduction
 from boomer.common.head_refinement cimport HeadRefinement
-from boomer.common.statistics cimport LabelMatrix
+from boomer.common.input_data cimport LabelMatrix
 from boomer.common.pruning cimport Pruning
 from boomer.common.post_processing cimport PostProcessor
 from boomer.common.sub_sampling cimport InstanceSubSampling, FeatureSubSampling, LabelSubSampling
