+++ python/boomer/seco/cpp/label_wise_statistics.cpp
@@ -8,7 +8,7 @@ using namespace statistics;
 
 LabelWiseRefinementSearchImpl::LabelWiseRefinementSearchImpl(
         rule_evaluation::LabelWiseRuleEvaluationImpl* ruleEvaluation, intp numPredictions, const intp* labelIndices,
-        AbstractLabelMatrix* labelMatrix, const float64* uncoveredLabels, const uint8* minorityLabels,
+        input::AbstractLabelMatrix* labelMatrix, const float64* uncoveredLabels, const uint8* minorityLabels,
         const float64* confusionMatricesTotal, const float64* confusionMatricesSubset) {
     ruleEvaluation_ = ruleEvaluation;
     numPredictions_ = numPredictions;
